+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add the 'images-folder' tag into module.xml file to deploy the modules images
 - Add the 'module:list' command, that shows the modules state 
 - Update Admin Logs to add the resource ID when available.
+- Add render smarty function, that executes the controller given in the action parameter.
 
 # 2.1.2